refacto steps
