+++ apps/site/app/scanner/event/[eventId]/scan/scanner.tsx
@@ -1,56 +1,26 @@
 "use client";
 
-import { useReducer, useState } from "react";
+import { useReducer } from "react";
 import { useToast } from "@/hooks/use-toast";
-import { Camera, PlusCircle, SaveIcon } from "lucide-react";
-import ReactModal from "react-modal";
+import { SaveIcon } from "lucide-react";
 
 import { sdk } from "@/lib/sdk";
-import { QrReader } from "@/components/qr-reader";
 import { buttonVariants } from "@/components/ui/button";
 import { reducer } from "../../../../../lib/utils_reducer";
+import { Cancel } from "./steps/Cancel";
+import { DisplayingResults } from "./steps/DisplayingResults";
 import { ManuallyEnteringPannel } from "./steps/ManuallyEnteringPannel";
 import { ManuallyEnteringTicket } from "./steps/ManuallyEnteringTicket";
+import { ScanReader } from "./steps/ScanReader";
+import { ScanningTicket } from "./steps/ScanningTicket";
+import { Start } from "./steps/Start";
 
 export const Scanner = () => {
-  const closeModal = () => {
-    setIsOpen(false);
-  };
-
   const [state, dispatch] = useReducer(reducer, {
     step: "start",
   });
   const { toast } = useToast();
-
-  const [resultModalIsOpen, setIsOpen] = useState<boolean>(true);
-
-  const [manualEmail, setManualEmail] = useState<string>();
-  const [errorEmail, setErrorEmail] = useState(false);
-  const customStyles = {
-    content: {
-      top: "50%",
-      left: "50%",
-      right: "auto",
-      bottom: "auto",
-      marginRight: "-40%",
-      transform: "translate(-50%, -50%)",
-      borderRadius: "16px",
-    },
-  };
-
   console.log("state", state);
-
-  const scanAttendee = () =>
-    sdk().ScanAttendee({
-      input: {
-        ticketPayload: {
-          ...state.ticket,
-          panelNumber: state.pannel,
-          email: state.email,
-        },
-      },
-    });
-  const emailPattern = /^\S+@\S+$/i;
 
   let offlineData = [];
 
@@ -96,7 +66,6 @@ export const Scanner = () => {
                     title: "✅ Synchronisation ok",
                   });
                 })
-
                 .catch((error) => {
                   console.log("error", error);
                   toast({
@@ -110,34 +79,9 @@ export const Scanner = () => {
         ) : null}
       </div>
       {state.step === "start" ? (
-        <div className="flex flex-col items-center justify-center">
-          <button
-            className={buttonVariants({ size: "sm" })}
-            onClick={() => {
-              dispatch({
-                type: "start_scanner",
-              });
-            }}
-          >
-            <Camera className="mr-2" /> Commencer à scanner
-          </button>
-        </div>
+        <Start state={state} dispatch={dispatch} />
       ) : state.step === "scanning_ticket" ? (
-        <>
-          <button
-            className={buttonVariants({ size: "sm" })}
-            onClick={() => {
-              dispatch({
-                type: "scan_ticket_error",
-                payload: {
-                  error: "le ticket n'est pas détecté",
-                },
-              });
-            }}
-          >
-            trigger error ticket
-          </button>
-        </>
+        <ScanningTicket state={state} dispatch={dispatch} />
       ) : state.step === "scanning_ticket_success" ? (
         <div className="flex flex-col items-center justify-center mx-auto">
           <button
@@ -186,166 +130,16 @@ export const Scanner = () => {
       ) : state.step === "manually_entering_pannel" ? (
         <ManuallyEnteringPannel state={state} dispatch={dispatch} />
       ) : state.step === "displaying_result" ? (
-        <>
-          <ReactModal isOpen={resultModalIsOpen} style={customStyles} onRequestClose={closeModal}>
-            <div className="flex flex-col p-2">
-              <h1 className="font-semibold text-center">Récapitulatif du scanning</h1>
-              {!state?.ticket?.email ? (
-                <p className="mb-4 text-xs text-red-600">Aucun email renseigné, souhaitez-vous en ajouter un ?</p>
-              ) : (
-                ""
-              )}
-              <span>Nom : {state?.ticket?.lastname}</span>
-              <span>Prénom : {state.ticket?.firstname}</span>
-              <span className="flex items-center">
-                Email :{" "}
-                {!state?.ticket?.email ? (
-                  <div className="flex items-center">
-                    <form
-                      onSubmit={() =>
-                        dispatch({
-                          type: "display_result",
-                          payload: {
-                            email: manualEmail,
-                          },
-                        })
-                      }
-                    >
-                      {" "}
-                      <input
-                        type="email"
-                        className="p-1 ml-2 border rounded-md"
-                        value={manualEmail}
-                        onChange={(e) => setManualEmail(e.target.value)}
-                      />{" "}
-                      {errorEmail === true ? <p className="text-red-600">Mail invalide</p> : ""}
-                    </form>
-
-                    <button
-                      type="submit"
-                      onClick={() => {
-                        console.log("+++++", state?.ticket?.attendeeId);
-                        manualEmail.match(emailPattern)
-                          ? dispatch({
-                              type: "display_result",
-                              payload: {
-                                email: manualEmail,
-                              },
-                            })
-                          : setErrorEmail(true);
-                      }}
-                    >
-                      <PlusCircle className="ml-2" />
-                    </button>
-                  </div>
-                ) : (
-                  state?.ticket?.email
-                )}
-              </span>
-              <span>Panneau : {state?.pannel} </span>
-              <div className="flex flex-col items-center">
-                <button
-                  type="button"
-                  className={buttonVariants({ size: "sm", className: "mt-12 bg-green-700 w-6/12 " })}
-                  onClick={() => {
-                    scanAttendee()
-                      .then(() => closeModal())
-                      .then((result) => {
-                        console.log("result", result);
-                        toast({
-                          title: "✅ Scan ok",
-                          description: "Participation scannée avec succès",
-                        });
-                      })
-
-                      .catch((error) => {
-                        console.log("error", error);
-                        toast({
-                          title: "⛔️ L'enregistrement a échoué",
-                          description: "Vous pourrez synchroniser plus tard",
-                        });
-                        dispatch({
-                          type: "synchronize",
-                          payload: {
-                            error:
-                              "L'enregistrement n'a pas fonctionné, les informations vont être stockées localement",
-                          },
-                        });
-                        console.error(error);
-                      })
-                      .then(() => {
-                        localStorage.setItem(
-                          "offlineData",
-                          JSON.stringify([...JSON.parse(localStorage.getItem("offlineData") || "[]"), state.ticket])
-                        ),
-                          console.log("synched in storage");
-                      })
-                      .then(() =>
-                        dispatch({
-                          type: "start_scanner",
-                        })
-                      );
-                  }}
-                >
-                  Valider
-                </button>
-              </div>
-            </div>
-          </ReactModal>
-        </>
+        <DisplayingResults state={state} dispatch={dispatch} />
       ) : state.step === "synchronizing" ? (
         <button type="button">Synchroniser</button>
       ) : null}
       <div className="flex flex-col items-center justify-center">
-        <button
-          className="p-2 m-2 text-sm text-white bg-red-700 border-none rounded-md"
-          onClick={() => {
-            dispatch({
-              type: "cancel",
-            });
-          }}
-        >
-          Annuler
-        </button>
+        <Cancel dispatch={dispatch} />
       </div>
-
       <div>
-        <QrReader
-          onResult={(result, error) => {
-            if (state.step === "scanning_ticket") {
-              console.log(result);
-              if (!!result) {
-                dispatch({
-                  type: "scan_ticket",
-                  payload: {
-                    ticket: JSON.parse(result.getText()),
-                  },
-                });
-                console.log(state.ticket);
-              }
-
-              if (!!error) {
-                console.log(error);
-              }
-            } else if (state.step === "scanning_pannel") {
-              if (!!result) {
-                dispatch({
-                  type: "scan_pannel",
-                  payload: {
-                    pannel: parseInt(result.getText()),
-                  },
-                });
-              }
-              if (!!error) {
-                console.log(error);
-              }
-            }
-          }}
-          className="flex flex-col w-6/12 mx-auto"
-          constraints={{}}
-        />
+        <ScanReader state={state} dispatch={dispatch} />
       </div>
-
       <pre>{JSON.stringify(state, null, 2)}</pre>
     </div>
   );
